Extract ha-media-upload-button
